+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 4.0.17
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   plate@6.3.9
+
 ## 4.0.16
 
 ### Patch Changes