+++ packages/amount-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 7.2.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@11.1.6
+    -   with-suffix@4.0.21
+
 ## 7.2.6
 
 ### Patch Changes