+++ packages/phone-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 7.1.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.1.6
+
 ## 7.1.5
 
 ### Patch Changes