+++ packages/slider-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 7.1.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@11.1.6
+
 ## 7.1.7
 
 ### Patch Changes