+++ packages/with-suffix/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 4.0.21
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@11.1.6
+
 ## 4.0.20
 
 ### Patch Changes